spanner/requirements.txt
google-cloud-spanner==3.49.1

